Lesson 19
Topic: FUNCTION
WHAT IS A FUNCTION?
A function is a set of codes designed to perform a specific task. So far, we have used several ready-made functions. For example, the print() function is used to output text to the console, and the range() function is used to generate numbers in a certain range.

In fact, there is several lines of code behind any function, but when we call the function, we only write its name. The code behind the function remains hidden from us. That is the convenience of functions. In order not to write certain codes repeatedly throughout the program, we can collect them, place them in one function, and call these codes throughout the program by the function name.

Functions are different, some functions accept a value from you and output some information to the console, while some perform various operations on the value received from you and return a new value. There are also functions that do not accept a value from the user at all.

In this topic, you will learn how to create, call, test, and debug new functions in Python. At the end of our lesson, we will also learn how to divide our program into several functions and place functions in separate files called modules.